# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v2/registry/categories:
        get:
            tags:
                - RegistryService
            operationId: RegistryService_ListCategories
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListCategoriesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /v2/registry/plugins:
        get:
            tags:
                - RegistryService
            operationId: RegistryService_ListPlugins
            parameters:
                - name: category
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListPluginsResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v2/registry/plugins/{pluginId}:
        get:
            tags:
                - RegistryService
            operationId: RegistryService_GetPlugin
            parameters:
                - name: pluginId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PluginManifest'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /v2/registry/sources:
        get:
            tags:
                - RegistryService
            operationId: RegistryService_ListSources
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListSourcesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
    /v2/users/{userId}/pipelines:
        get:
            tags:
                - PipelineService
            operationId: PipelineService_ListPipelines
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListPipelinesResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
        post:
            tags:
                - PipelineService
            operationId: PipelineService_CreatePipeline
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PipelineConfig'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PipelineConfig'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        ConfigFieldDependency:
            type: object
            properties:
                fieldKey:
                    type: string
                values:
                    type: array
                    items:
                        type: string
        ConfigFieldOption:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
        ConfigFieldSchema:
            type: object
            properties:
                key:
                    type: string
                label:
                    type: string
                description:
                    type: string
                fieldType:
                    type: integer
                    format: enum
                required:
                    type: boolean
                defaultValue:
                    type: string
                options:
                    type: array
                    items:
                        $ref: '#/components/schemas/ConfigFieldOption'
                validation:
                    $ref: '#/components/schemas/ConfigFieldValidation'
                dependsOn:
                    $ref: '#/components/schemas/ConfigFieldDependency'
                keyOptions:
                    type: array
                    items:
                        $ref: '#/components/schemas/ConfigFieldOption'
                valueOptions:
                    type: array
                    items:
                        $ref: '#/components/schemas/ConfigFieldOption'
                dynamicSource:
                    type: string
                valueDynamicSource:
                    type: string
        ConfigFieldValidation:
            type: object
            properties:
                minLength:
                    type: integer
                    format: int32
                maxLength:
                    type: integer
                    format: int32
                pattern:
                    type: string
                minValue:
                    type: number
                    format: double
                maxValue:
                    type: number
                    format: double
...
        DestinationConfig:
            type: object
            properties:
                config:
                    type: object
                    additionalProperties:
                        type: string
                excludedEnrichers:
                    type: array
                    items:
                        type: string
...
        EnricherConfig:
            type: object
            properties:
                providerType:
                    type: integer
                    format: enum
                typedConfig:
                    type: object
                    additionalProperties:
                        type: string
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        ListCategoriesResponse:
            type: object
            properties:
                categories:
                    type: array
                    items:
                        type: string
...
        ListPipelinesResponse:
            type: object
            properties:
                pipelines:
                    type: array
                    items:
                        $ref: '#/components/schemas/PipelineConfig'
        ListPluginsResponse:
            type: object
            properties:
                plugins:
                    type: array
                    items:
                        $ref: '#/components/schemas/PluginManifest'
...
        ListSourcesResponse:
            type: object
            properties:
                sources:
                    type: array
                    items:
                        $ref: '#/components/schemas/PluginManifest'
...
        PipelineConfig:
            type: object
            properties:
                id:
                    type: string
                source:
                    type: string
                enrichers:
                    type: array
                    items:
                        $ref: '#/components/schemas/EnricherConfig'
                destinations:
                    type: array
                    items:
                        type: integer
                        format: enum
                name:
                    type: string
                disabled:
                    type: boolean
                sourceConfig:
                    type: object
                    additionalProperties:
                        type: string
                destinationConfigs:
                    type: object
                    additionalProperties:
                        $ref: '#/components/schemas/DestinationConfig'
...
        PluginManifest:
            type: object
            properties:
                id:
                    type: string
                type:
                    type: integer
                    format: enum
                name:
                    type: string
                description:
                    type: string
                icon:
                    type: string
                configSchema:
                    type: array
                    items:
                        $ref: '#/components/schemas/ConfigFieldSchema'
                requiredIntegrations:
                    type: array
                    items:
                        type: string
                enabled:
                    type: boolean
                enricherProviderType:
                    type: integer
                    format: int32
                destinationType:
                    type: integer
                    format: int32
                marketingDescription:
                    type: string
                features:
                    type: array
                    items:
                        type: string
                transformations:
                    type: array
                    items:
                        $ref: '#/components/schemas/Transformation'
                useCases:
                    type: array
                    items:
                        type: string
                externalUrlTemplate:
                    type: string
                requiredTier:
                    type: string
                category:
                    type: string
                sortOrder:
                    type: integer
                    format: int32
                isPremium:
                    type: boolean
                popularityScore:
                    type: integer
                    format: int32
                iconType:
                    type: string
                iconPath:
                    type: string
                isTemporarilyUnavailable:
                    type: boolean
                allowMultipleInstances:
                    type: boolean
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        Transformation:
            type: object
            properties:
                field:
                    type: string
                label:
                    type: string
                before:
                    type: string
                after:
                    type: string
                visualType:
                    type: string
                afterHtml:
                    type: string
...