# src/go/services/api-public/internal/server/server.go
package server

import (
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"

	"github.com/fitglue/server/src/go/internal/infra"
	activitypb "github.com/fitglue/server/src/go/pkg/types/pb/services/activity"
	registrypb "github.com/fitglue/server/src/go/pkg/types/pb/services/registry"
)

// APIServer implements the HTTP router interfacing with FitGlue domain gRPC services
type APIServer struct {
	router      *chi.Mux
	logger      infra.Logger
	activitySvc activitypb.ActivityServiceClient
	registrySvc registrypb.RegistryServiceClient
}

// NewAPIServer constructs the application routing and API middleware stack
func NewAPIServer(
	logger infra.Logger,
	activitySvc activitypb.ActivityServiceClient,
	registrySvc registrypb.RegistryServiceClient,
) *APIServer {
	s := &APIServer{
		router:      chi.NewRouter(),
		logger:      logger,
		activitySvc: activitySvc,
		registrySvc: registrySvc,
	}

	s.setupRoutes()
	return s
}

// ServeHTTP implements http.Handler automatically so the APIServer can be bound to net/http
func (s *APIServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	s.router.ServeHTTP(w, r)
}

func (s *APIServer) setupRoutes() {
	// Root level middleware (shared by all endpoints)
	s.router.Use(middleware.RequestID)
	s.router.Use(middleware.RealIP)
	s.router.Use(middleware.Recoverer)

	// API Public block (No Auth / API routing)
	s.router.Route("/api/public", func(r chi.Router) {
		r.Get("/health", func(w http.ResponseWriter, r *http.Request) {
			w.WriteHeader(http.StatusOK)
		})

		s.registerRegistryRoutes(r)
		s.registerShowcaseRoutes(r)
	})
}

func (s *APIServer) registerRegistryRoutes(r chi.Router) {
	r.Get("/registry/plugins", s.handleListPlugins)
	r.Get("/registry/plugins/{id}", s.handleGetPlugin)
	r.Get("/registry/categories", s.handleListCategories)
	r.Get("/registry/sources", s.handleListSources)

	// Wait, is endpoints mapping to API Public list Destinations requested?
	// Let's implement handles.
}

func (s *APIServer) registerShowcaseRoutes(r chi.Router) {
	r.Get("/showcase/{id}", s.handleGetPublicShowcase)
}

func (s *APIServer) handleListPlugins(w http.ResponseWriter, r *http.Request) {
	req := &registrypb.ListPluginsRequest{
		Category: r.URL.Query().Get("category"),
	}

	res, err := s.registrySvc.ListPlugins(r.Context(), req)
	if err != nil {
		WriteError(w, err)
		return
	}

	WriteJSON(w, res)
}

func (s *APIServer) handleGetPlugin(w http.ResponseWriter, r *http.Request) {
	req := &registrypb.GetPluginRequest{
		PluginId: chi.URLParam(r, "id"),
	}

	res, err := s.registrySvc.GetPlugin(r.Context(), req)
	if err != nil {
		WriteError(w, err)
		return
	}

	WriteJSON(w, res)
}

func (s *APIServer) handleListCategories(w http.ResponseWriter, r *http.Request) {
	req := &registrypb.ListCategoriesRequest{}

	res, err := s.registrySvc.ListCategories(r.Context(), req)
	if err != nil {
		WriteError(w, err)
		return
	}

	WriteJSON(w, res)
}

func (s *APIServer) handleListSources(w http.ResponseWriter, r *http.Request) {
	req := &registrypb.ListSourcesRequest{}

	res, err := s.registrySvc.ListSources(r.Context(), req)
	if err != nil {
		WriteError(w, err)
		return
	}

	WriteJSON(w, res)
}

func (s *APIServer) handleGetPublicShowcase(w http.ResponseWriter, r *http.Request) {
	req := &activitypb.GetPublicShowcaseRequest{
		ShowcaseId: chi.URLParam(r, "id"),
	}

	res, err := s.activitySvc.GetPublicShowcase(r.Context(), req)
	if err != nil {
		WriteError(w, err)
		return
	}

	WriteJSON(w, res)
}

// statusError is a helper for manually generating an error satisfying gRPC status layout
func statusError(code int, msg string) error {
	// Simple wrapper for non-gRPC errors to use WriteError
	return &CustomError{HTTPCode: code, Msg: msg}
}

// CustomError helps map generic HTTP errors into our WriteError handler
type CustomError struct {
	HTTPCode int
	Msg      string
}

func (e *CustomError) Error() string {
	return e.Msg
}